+++ stocks/admin.py
@@ -1,7 +1,8 @@
 from django.contrib import admin
 
 from django.contrib import admin
-from .models import StockStore
+from .models.stock_model import StockStore
+
 
 # Register your models here.
-admin.site.register(StockStore)
+admin.site.register([StockStore])
