+++ stocks/fixtures/fixtures.json
